Added Percona 5.7 XtraDB cluster.
I encountered problems with DTrace support. Apparantly the generated
header files are incorrect so support is disabled for now.

Unlike the percona-cluster56 package Sphinx support was removed from
this one. The percona57-server doesn't include it either. The reason for
this is the issue described here:
http://sphinxsearch.com/forum/view.html?id=13908

libwrap support was disabled due to this issue:
https://bugs.mysql.com/bug.php?id=84495 . Apparantly the fix hasn't
found it's way in upstream Percona XtraDB yet.