gcc9: Add m4 to USE_TOOLS.
This is required due to complicated reasons.  Flex hardcodes the
location to GNU m4 in its build and depends on it, but as the GCC build
sets M4=m4 in its environment, and because the pkgsrc build has a
limited $PATH, it ends up trying to use the native m4 which lacks -P.

Adding m4 to tools, and setting TOOLS_PLATFORM.m4 for the gcc build in
exttools.mk ensures that at least M4=m4, when searching by $PATH, finds
the GNU m4 pointed at by tools.